chore: clean seed corpus
