+++ .github/workflows/cflite-cron.yml
@@ -62,5 +62,5 @@ jobs:
         uses: google/clusterfuzzlite/actions/run_fuzzers@884713a6c30a92e5e8544c39945cd7cb630abcd1 #v1
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
-          fuzz-seconds: 300
+          fuzz-seconds: 600
           mode: coverage
