+++ .github/workflows/cflite-pr.yml
@@ -53,7 +53,7 @@ jobs:
         uses: google/clusterfuzzlite/actions/run_fuzzers@884713a6c30a92e5e8544c39945cd7cb630abcd1 #v1
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
-          fuzz-seconds: 300
+          fuzz-seconds: 600
           mode: code-change
           sanitizer: ${{ matrix.sanitizer }}
           output-sarif: true
