+++ internal/pkg/merge/fuzz.go
@@ -1,6 +1,6 @@
 // -----------------------------------------------------------------------------
 // SPDX-FileCopyrightText: Copyright © 2024 bomctl a Series of LF Projects, LLC
-// SPDX-FileName: internal/pkg/fetch/fetch_fuzz.go
+// SPDX-FileName: internal/pkg/merge/fuzz.go
 // SPDX-FileType: SOURCE
 // SPDX-License-Identifier: Apache-2.0
 // -----------------------------------------------------------------------------
@@ -17,7 +17,10 @@
 // limitations under the License.
 // -----------------------------------------------------------------------------
 
-package fetch
+//go:build fuzz
+// +build fuzz
+
+package merge
 
 import (
 	"context"
@@ -26,15 +29,11 @@ import (
 	"github.com/bomctl/bomctl/internal/pkg/options"
 )
 
-const testURL = "https://raw.githubusercontent.com/bomctl/bomctl-playground/main/bomctl_bomctl_v0.3.0.cdx.json"
-
-func FuzzFetch(f *testing.F) {
-	f.Add([]byte(testURL))
-
-	f.Fuzz(func(t *testing.T, data []byte) {
-		_, err := Fetch(
-			string(data),
-			&options.FetchOptions{Options: options.New().WithContext(context.Background())},
+func FuzzMerge(f *testing.F) {
+	f.Fuzz(func(t *testing.T, id1, id2 string) {
+		_, err := Merge(
+			[]string{id1, id2},
+			&options.MergeOptions{Options: options.New().WithContext(context.Background())},
 		)
 
 		if err == nil {
